requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
asyncpg==0.29.0
aiosqlite==0.20.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
websockets==12.0
httpx==0.27.0
greenlet==3.1.1
aiosmtplib==3.0.1
